+++ packages/core/src/components/widgets/CheckboxWidget.tsx
@@ -9,6 +9,7 @@ import {
   RJSFSchema,
   StrictRJSFSchema,
   WidgetProps,
+  getUiOptions,
 } from '@rjsf/utils';
 
 /** The `CheckBoxWidget` is a widget for rendering boolean properties.
@@ -57,8 +58,10 @@ function CheckboxWidget<T = any, S extends StrictRJSFSchema = RJSFSchema, F exte
     (event: FocusEvent<HTMLInputElement>) => onFocus(id, event.target.checked),
     [onFocus, id],
   );
-  const description = options.description ?? schema.description;
 
+  const uiOptions = getUiOptions(uiSchema);
+  const isCheckboxWidget = uiOptions.widget === 'checkbox';
+  const description = isCheckboxWidget ? undefined : (options.description ?? schema.description);
   return (
     <div className={`checkbox ${disabled || readonly ? 'disabled' : ''}`}>
       {!hideLabel && description && (